use pull instead of reset.
Don't know if something changed in git defaults, but the reset was leaving the branch with a modified file, instead of updating it to the latest version. Causing a lot of weirdness: basically the branches didn't get swapped, and since there was modification in the base branch, `git checkout ` failled because the tree was dirty.